requirements.txt
python-telegram-bot
schedule
python-dotenv
sqlalchemy
